+++ src/Config.php
@@ -3,8 +3,15 @@ declare(strict_types = 1);
 
 namespace Innmind\OperatingSystem;
 
+use Innmind\TimeContinuum\{
+    Clock,
+    Earth,
+    ElapsedPeriod,
+};
 use Innmind\Filesystem\CaseSensitivity;
 use Innmind\Server\Status\EnvironmentPath;
+use Innmind\TimeWarp\Halt;
+use Innmind\IO\IO;
 use Innmind\Stream\{
     Capabilities,
     Streams,
@@ -13,8 +20,11 @@ use Innmind\Immutable\Maybe;
 
 final class Config
 {
+    private Clock $clock;
     private CaseSensitivity $caseSensitivity;
     private Capabilities $streamCapabilities;
+    private IO $io;
+    private Halt $halt;
     private EnvironmentPath $path;
     /** @var Maybe<positive-int> */
     private Maybe $maxHttpConcurrency;
@@ -23,13 +33,19 @@ final class Config
      * @param Maybe<positive-int> $maxHttpConcurrency
      */
     private function __construct(
+        Clock $clock,
         CaseSensitivity $caseSensitivity,
         Capabilities $streamCapabilities,
+        IO $io,
+        Halt $halt,
         EnvironmentPath $path,
         Maybe $maxHttpConcurrency,
     ) {
+        $this->clock = $clock;
         $this->caseSensitivity = $caseSensitivity;
         $this->streamCapabilities = $streamCapabilities;
+        $this->io = $io;
+        $this->halt = $halt;
         $this->path = $path;
         $this->maxHttpConcurrency = $maxHttpConcurrency;
     }
@@ -40,10 +56,32 @@ final class Config
         $maxHttpConcurrency = Maybe::nothing();
 
         return new self(
+            new Earth\Clock,
             CaseSensitivity::sensitive,
-            Streams::fromAmbientAuthority(),
+            $streams = Streams::fromAmbientAuthority(),
+            IO::of(static fn(?ElapsedPeriod $timeout) => match ($timeout) {
+                null => $streams->watch()->waitForever(),
+                default => $streams->watch()->timeoutAfter($timeout),
+            }),
+            new Halt\Usleep,
             EnvironmentPath::of(\getenv('PATH') ?: ''),
             $maxHttpConcurrency,
+        );
+    }
+
+    /**
+     * @psalm-mutation-free
+     */
+    public function withClock(Clock $clock): self
+    {
+        return new self(
+            $clock,
+            $this->caseSensitivity,
+            $this->streamCapabilities,
+            $this->io,
+            $this->halt,
+            $this->path,
+            $this->maxHttpConcurrency,
         );
     }
 
@@ -53,8 +91,11 @@ final class Config
     public function caseInsensitiveFilesystem(): self
     {
         return new self(
+            $this->clock,
             CaseSensitivity::insensitive,
             $this->streamCapabilities,
+            $this->io,
+            $this->halt,
             $this->path,
             $this->maxHttpConcurrency,
         );
@@ -65,9 +106,32 @@ final class Config
      */
     public function useStreamCapabilities(Capabilities $capabilities): self
     {
+        /** @psalm-suppress ImpureMethodCall This should be solved in the next innmind/io release */
         return new self(
+            $this->clock,
             $this->caseSensitivity,
             $capabilities,
+            IO::of(static fn(?ElapsedPeriod $timeout) => match ($timeout) {
+                null => $capabilities->watch()->waitForever(),
+                default => $capabilities->watch()->timeoutAfter($timeout),
+            }),
+            $this->halt,
+            $this->path,
+            $this->maxHttpConcurrency,
+        );
+    }
+
+    /**
+     * @psalm-mutation-free
+     */
+    public function haltProcessVia(Halt $halt): self
+    {
+        return new self(
+            $this->clock,
+            $this->caseSensitivity,
+            $this->streamCapabilities,
+            $this->io,
+            $halt,
             $this->path,
             $this->maxHttpConcurrency,
         );
@@ -79,8 +143,11 @@ final class Config
     public function withEnvironmentPath(EnvironmentPath $path): self
     {
         return new self(
+            $this->clock,
             $this->caseSensitivity,
             $this->streamCapabilities,
+            $this->io,
+            $this->halt,
             $path,
             $this->maxHttpConcurrency,
         );
@@ -94,11 +161,22 @@ final class Config
     public function limitHttpConcurrencyTo(int $max): self
     {
         return new self(
+            $this->clock,
             $this->caseSensitivity,
             $this->streamCapabilities,
+            $this->io,
+            $this->halt,
             $this->path,
             Maybe::just($max),
         );
+    }
+
+    /**
+     * @internal
+     */
+    public function clock(): Clock
+    {
+        return $this->clock;
     }
 
     /**
@@ -115,6 +193,22 @@ final class Config
     public function streamCapabilities(): Capabilities
     {
         return $this->streamCapabilities;
+    }
+
+    /**
+     * @internal
+     */
+    public function io(): IO
+    {
+        return $this->io;
+    }
+
+    /**
+     * @internal
+     */
+    public function halt(): Halt
+    {
+        return $this->halt;
     }
 
     /**