+++ src/OperatingSystem/Unix.php
@@ -21,11 +21,9 @@ use Innmind\Server\Control\{
     Servers,
 };
 use Innmind\TimeContinuum\Clock;
-use Innmind\TimeWarp\Halt\Usleep;
 
 final class Unix implements OperatingSystem
 {
-    private Clock $clock;
     private Config $config;
     private ?Filesystem $filesystem = null;
     private ?ServerStatus $status = null;
@@ -35,36 +33,30 @@ final class Unix implements OperatingSystem
     private ?Remote $remote = null;
     private ?CurrentProcess $process = null;
 
-    private function __construct(Clock $clock, Config $config)
+    private function __construct(Config $config)
     {
-        $this->clock = $clock;
         $this->config = $config;
     }
 
-    public static function of(Clock $clock, Config $config = null): self
+    public static function of(Config $config = null): self
     {
-        return new self($clock, $config ?? Config::of());
+        return new self($config ?? Config::of());
     }
 
-    /**
-     * @internal
-     */
-    public function config(): Config
+    public function map(callable $map): OperatingSystem
     {
-        return $this->config;
+        return $map($this, $this->config);
     }
 
     public function clock(): Clock
     {
-        return $this->clock;
+        return $this->config->clock();
     }
 
     public function filesystem(): Filesystem
     {
         return $this->filesystem ??= Filesystem\Generic::of(
             $this->control()->processes(),
-            new Usleep,
-            $this->clock,
             $this->config,
         );
     }
@@ -83,7 +75,7 @@ final class Unix implements OperatingSystem
         return $this->control ??= Servers\Unix::of(
             $this->clock(),
             $this->config->streamCapabilities(),
-            new Usleep,
+            $this->config->halt(),
         );
     }
 
@@ -101,13 +93,12 @@ final class Unix implements OperatingSystem
     {
         return $this->remote ??= Remote\Generic::of(
             $this->control(),
-            $this->clock(),
             $this->config,
         );
     }
 
     public function process(): CurrentProcess
     {
-        return $this->process ??= CurrentProcess\Generic::of(new Usleep);
+        return $this->process ??= CurrentProcess\Generic::of($this->config->halt());
     }
 }