+++ tests/Filesystem/GenericTest.php
@@ -6,12 +6,11 @@ namespace Tests\Innmind\OperatingSystem\Filesystem;
 use Innmind\OperatingSystem\{
     Filesystem\Generic,
     Filesystem,
+    Config,
 };
 use Innmind\Filesystem\Adapter\Filesystem as FilesystemAdapter;
 use Innmind\Url\Path;
 use Innmind\Server\Control\Server\Processes;
-use Innmind\TimeWarp\Halt;
-use Innmind\TimeContinuum\Clock;
 use Innmind\FileWatch\Ping;
 use Fixtures\Innmind\Url\Path as FPath;
 use PHPUnit\Framework\TestCase;
@@ -27,8 +26,7 @@ class GenericTest extends TestCase
             Filesystem::class,
             Generic::of(
                 $this->createMock(Processes::class),
-                $this->createMock(Halt::class),
-                $this->createMock(Clock::class),
+                Config::of(),
             ),
         );
     }
@@ -37,8 +35,7 @@ class GenericTest extends TestCase
     {
         $filesystem = Generic::of(
             $this->createMock(Processes::class),
-            $this->createMock(Halt::class),
-            $this->createMock(Clock::class),
+            Config::of(),
         );
 
         $adapter = $filesystem->mount(Path::of('/tmp/'));
@@ -50,8 +47,7 @@ class GenericTest extends TestCase
     {
         $filesystem = Generic::of(
             $this->createMock(Processes::class),
-            $this->createMock(Halt::class),
-            $this->createMock(Clock::class),
+            Config::of(),
         );
 
         $adapter = $filesystem->mount(Path::of('/tmp/'));
@@ -63,8 +59,7 @@ class GenericTest extends TestCase
     {
         $filesystem = Generic::of(
             $this->createMock(Processes::class),
-            $this->createMock(Halt::class),
-            $this->createMock(Clock::class),
+            Config::of(),
         );
 
         $this->assertFalse($filesystem->contains(Path::of('/tmp/foo')));
@@ -77,8 +72,7 @@ class GenericTest extends TestCase
     {
         $filesystem = Generic::of(
             $this->createMock(Processes::class),
-            $this->createMock(Halt::class),
-            $this->createMock(Clock::class),
+            Config::of(),
         );
 
         $this->assertFalse($filesystem->contains(Path::of('/tmp/some-dir/')));
@@ -91,8 +85,7 @@ class GenericTest extends TestCase
     {
         $filesystem = Generic::of(
             $this->createMock(Processes::class),
-            $this->createMock(Halt::class),
-            $this->createMock(Clock::class),
+            Config::of(),
         );
 
         $this->assertInstanceOf(Ping::class, $filesystem->watch(Path::of('/somewhere')));
@@ -105,8 +98,7 @@ class GenericTest extends TestCase
             ->then(function($path) {
                 $filesystem = Generic::of(
                     $this->createMock(Processes::class),
-                    $this->createMock(Halt::class),
-                    $this->createMock(Clock::class),
+                    Config::of(),
                 );
 
                 $this->assertFalse($filesystem->require($path)->match(
@@ -120,8 +112,7 @@ class GenericTest extends TestCase
     {
         $filesystem = Generic::of(
             $this->createMock(Processes::class),
-            $this->createMock(Halt::class),
-            $this->createMock(Clock::class),
+            Config::of(),
         );
 
         $this->assertSame(42, $filesystem->require(Path::of(__DIR__.'/fixture.php'))->match(