+++ tests/Remote/GenericTest.php
@@ -6,6 +6,7 @@ namespace Tests\Innmind\OperatingSystem\Remote;
 use Innmind\OperatingSystem\{
     Remote\Generic,
     Remote,
+    Config,
 };
 use Innmind\Server\Control\{
     Server,
@@ -16,7 +17,6 @@ use Innmind\Url\{
     Url,
     Authority\Port,
 };
-use Innmind\TimeContinuum\Clock;
 use Innmind\Socket\{
     Internet\Transport,
     Client\Internet,
@@ -39,7 +39,7 @@ class GenericTest extends TestCase
             Remote::class,
             Generic::of(
                 $this->createMock(Server::class),
-                $this->createMock(Clock::class),
+                Config::of(),
             ),
         );
     }
@@ -48,7 +48,7 @@ class GenericTest extends TestCase
     {
         $remote = Generic::of(
             $server = $this->createMock(Server::class),
-            $this->createMock(Clock::class),
+            Config::of(),
         );
         $server
             ->expects($this->once())
@@ -71,7 +71,7 @@ class GenericTest extends TestCase
     {
         $remote = Generic::of(
             $server = $this->createMock(Server::class),
-            $this->createMock(Clock::class),
+            Config::of(),
         );
         $server
             ->expects($this->once())
@@ -94,7 +94,7 @@ class GenericTest extends TestCase
     {
         $remote = Generic::of(
             $this->createMock(Server::class),
-            $this->createMock(Clock::class),
+            Config::of(),
         );
         $server = InternetServer::of(Transport::tcp(), IPv4::localhost(), Port::of(1234))->match(
             static fn($server) => $server,
@@ -115,7 +115,7 @@ class GenericTest extends TestCase
     {
         $remote = Generic::of(
             $this->createMock(Server::class),
-            $this->createMock(Clock::class),
+            Config::of(),
         );
 
         $http = $remote->http();
@@ -131,7 +131,7 @@ class GenericTest extends TestCase
             ->then(function($server) {
                 $remote = Generic::of(
                     $this->createMock(Server::class),
-                    $this->createMock(Clock::class),
+                    Config::of(),
                 );
 
                 $sql = $remote->sql($server);